Safari-611.1.10/LayoutTests/accessibility/mac/element-paths-expected.txt


This tests SVG group elements are accessible and that the svg:title element is returned properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svg.isAttributeSupported('AXPath') is true
SVG path description
Start Path
	Move to point
	Line to
	Line to
	Close
	Move to point

PASS map1.isAttributeSupported('AXPath') is true
Map1 path description
Start Path
	Move to point
	Line to
	Line to
	Line to
	Line to
	Close
	Move to point

PASS map2.isAttributeSupported('AXPath') is true
Map2 path description
Start Path
	Move to point
	Curve to
	Curve to
	Curve to
	Curve to
	Close
	Move to point

PASS successfullyParsed is true

TEST COMPLETE
